Made a much better logo! Need to look at DEP next
A DEP occurs when we try to hot reload the application code while the
executable is running. This is likely because the code and redraw code
are called by timers, so the redraw function tries to call the dll code
while it is unloaded (set to 0/NULL) during the update, since the
violation happens at address 0.